pyproject.toml
[tool.poetry]
name = "testwork"
version = "0.1.0"
description = ""
authors = ["Efim Kashapov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.4.0"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
